+++ Controller/Payment/Callback/Dmn.php
@@ -224,7 +224,6 @@ class Dmn extends Action implements CsrfAwareActionInterface
             return $this->jsonOutput;
         }
         
-        // Here will be set $this->orderIncrementId or $this->quoteId
         $this->getOrderIdentificators();
         
         /**
@@ -1096,89 +1095,6 @@ class Dmn extends Action implements CsrfAwareActionInterface
     }
 
     /**
-     * Place order.
-     *
-     * @param array $params
-     */
-//    private function placeOrder($params)
-//    {
-//        $this->readerWriter->createLog($params, 'PlaceOrder()');
-//        
-//        $result = $this->dataObjectFactory->create();
-//        
-//        if (empty($params['quote'])) {
-//            return $result
-//                ->setData('error', true)
-//                ->setData('message', 'Missing Quote parameter.');
-//        }
-//        
-//        try {
-//            $quote = $this->quoteFactory->create()->loadByIdWithoutStore((int) $params['quote']);
-//            
-//            if (!is_object($quote)) {
-//                $this->readerWriter->createLog($quote, 'placeOrder error - the quote is not an object.');
-//
-//                return $result
-//                    ->setData('error', true)
-//                    ->setData('message', 'The quote is not an object.');
-//            }
-//            
-//            $method = $quote->getPayment()->getMethod();
-//            
-//            $this->readerWriter->createLog(
-//                [
-//                    'quote payment Method'  => $method,
-//                    'quote id'              => $quote->getEntityId(),
-//                    'quote is active'       => $quote->getIsActive(),
-////                    'quote reserved ord id' => $quote->getReservedOrderId(),
-//                ],
-//                'Quote data'
-//            );
-//
-//            if ((int) $quote->getIsActive() == 0) {
-//                $this->readerWriter->createLog($quote->getQuoteId(), 'Quote ID');
-//
-//                return $result
-//                    ->setData('error', true)
-//                    ->setData('message', 'Quote is not active.');
-//            }
-//
-//            if ($method !== Payment::METHOD_CODE) {
-//                return $result
-//                    ->setData('error', true)
-//                    ->setData('message', 'Quote payment method is "' . $method . '"');
-//            }
-//
-//            //            $params = array_merge(
-//            //                $this->request->getParams(),
-//            //                $this->request->getPostValue()
-//            //            );
-//            
-//            $orderId = $this->cartManagement->placeOrder($params);
-//
-//            $result
-//                ->setData('success', true)
-//                ->setData('order_id', $orderId);
-//
-//            $this->_eventManager->dispatch(
-//                'nuvei_place_order',
-//                [
-//                    'result' => $result,
-//                    'action' => $this,
-//                ]
-//            );
-//        } catch (\Exception $exception) {
-//            $this->readerWriter->createLog($exception->getMessage(), 'DMN placeOrder Exception: ');
-//            
-//            return $result
-//                ->setData('error', true)
-//                ->setData('message', $exception->getMessage());
-//        }
-//        
-//        return $result;
-//    }
-    
-    /**
      * Function validateChecksum
      *
      * @param array  $params
@@ -1443,13 +1359,16 @@ class Dmn extends Action implements CsrfAwareActionInterface
             '$identificator name'   => $field,
         ]);
         
+        // If the DMN is not Sale or Auth check just once. The Order must be already there.
+        if (!empty($this->params['transactionType'])
+            && !in_array($this->params['transactionType'], ['Sale', 'Auth'])
+        ) {
+            $this->loop_max_tries = 1;
+        }
+        
         try {
             $searchCriteria = $this->searchCriteriaBuilder
-                ->addFilter(
-                    $field,
-                    $value,
-                    'eq'
-                )->create();
+                ->addFilter($field, $value, 'eq')->create();
 
             do {
                 $this->loop_tries++;